+++ packages/expo-router/CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ### 🎉 New features
 
+- add icon api compatible with react-native-screens 4.16 and 4.18 ([#41631](https://github.com/expo/expo/pull/41631) by [@Ubax](https://github.com/Ubax))
+
 ### 🐛 Bug fixes
 
 - [web] Fix dynamic route params becoming query params in getPathFromState. ([#41488](https://github.com/expo/expo/pull/41488) by [@erickreutz](https://github.com/erickreutz))